Module melior_next::pass::conversion

Dialect conversion passes.

Functions

convert_affine_for_to_gpu

Creates a pass to convert the affine for dialect to the gpu dialect.

convert_affine_to_standard

Creates a pass to convert the affine dialect to the standard dialect.

convert_affiner_for_to_gpu

Creates a pass to convert the affiner for dialect to the gpu dialect.

convert_amdgpu_to_rocdl

convert_arith_to_llvm

convert_arith_to_spirv

convert_arithmetic_to_llvm

Creates a pass to convert the arith dialect to the llvm dialect.

convert_arm_neon2d_to_intr

convert_async_to_llvm

Creates a pass to convert the async dialect to the llvm dialect.

convert_bufferization_to_mem_ref

convert_cf_to_llvm

Creates a pass to convert the cf dialect to the llvm dialect.

convert_cf_to_spirv

Creates a pass to convert the cf dialect to the spirv dialect.

convert_complex_to_libm

convert_complex_to_llvm

convert_complex_to_standard

convert_control_flow_to_llvm

convert_control_flow_to_spirv

convert_func_to_llvm

Creates a pass to convert the func dialect to the llvm dialect.

convert_func_to_spirv

convert_gpu_launch_func_to_vulkan_launch_func

convert_gpu_ops_to_nvvmops

convert_gpu_ops_to_rocdlops

convert_gpu_to_llvm

Creates a pass to convert the gpu dialect to the llvm dialect.

convert_gpu_to_llvmconversion_pass

convert_gpu_to_spirv

convert_index_to_llvm

Creates a pass to convert the builtin index to the llvm dialect.

convert_index_to_llvmpass

convert_linalg_to_llvm

convert_linalg_to_standard

convert_lower_host_code_to_llvm

convert_map_memref_storage_class

convert_math_to_funcs

convert_math_to_libm

Creates a pass to convert the math dialect to the libm dialect.

convert_math_to_llvm

Creates a pass to convert the math dialect to the llvm dialect.

convert_math_to_spirv

Creates a pass to convert the math dialect to the spirv dialect.

convert_mem_ref_to_spirv

convert_memref_to_llvmconversion_pass

convert_nvgpu_to_nvvm

convert_open_acc_to_llvm

convert_open_acc_to_scf

convert_openmp_to_llvm

convert_parallel_loop_to_gpu

convert_pdl_to_pdlinterp

convert_reconcile_casts

Creates a pass to reconcile builtin.unrealized_conversion_cast

convert_reconcile_unrealized_casts

convert_scf_to_cf

Creates a pass to convert the scf dialect to the cf dialect.

convert_scf_to_openmp

convert_scf_to_spirv

convert_shape_constraints

convert_shape_to_standard

convert_spirv_to_llvm

convert_tensor_to_linalg

convert_tensor_to_spirv

convert_tosa_to_arith

convert_tosa_to_linalg

convert_tosa_to_linalg_named

convert_tosa_to_scf

convert_tosa_to_tensor

convert_vector_to_gpu

convert_vector_to_llvm

convert_vector_to_scf

convert_vector_to_spirv

convert_vulkan_launch_func_to_vulkan_calls